status enum to int

+++ src/OpenTelemetry.Instrumentation.Http/Implementation/HttpWebRequestActivitySource.netfx.cs
@@ -426,7 +426,7 @@ internal static class HttpWebRequestActivitySource
 
             if (httpStatusCode.HasValue)
             {
-                tags.Add(SemanticConventions.AttributeHttpStatusCode, httpStatusCode.Value);
+                tags.Add(SemanticConventions.AttributeHttpStatusCode, (int)httpStatusCode.Value);
             }
 
             HttpClientDuration.Record(durationMs, tags);
